+++ test/DataGenerator/Composer/HyphenatedVersionRangeGenerator.php
@@ -1,0 +1,42 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2017-2023 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/version-constraint
+ */
+
+namespace Ergebnis\VersionConstraint\Test\DataGenerator\Composer;
+
+use Ergebnis\VersionConstraint\Test;
+
+/**
+ * @see https://getcomposer.org/doc/articles/versions.md#hyphenated-version-range-
+ */
+final class HyphenatedVersionRangeGenerator implements Test\DataGenerator\StringGenerator
+{
+    private readonly ExactVersionConstraintGenerator $exactVersionConstraintGenerator;
+
+    public function __construct()
+    {
+        $this->exactVersionConstraintGenerator = new ExactVersionConstraintGenerator();
+    }
+
+    public function generate(): \Generator
+    {
+        foreach ($this->exactVersionConstraintGenerator->generate() as $leftExactVersionConstraint) {
+            foreach ($this->exactVersionConstraintGenerator->generate() as $rightExactVersionConstraint) {
+                yield \sprintf(
+                    '%s - %s',
+                    $leftExactVersionConstraint,
+                    $rightExactVersionConstraint,
+                );
+            }
+        }
+    }
+}