+++ CHANGELOG.md
@@ -8,4 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`dc5f5d1...main`][dc5f5d1...main].
 
+### Added
+
+- Added `Composer\TildeVersionRange` as a value object ([#2]), by [@localheinz]
+
 [dc5f5d1...main]: https://github.com/ergebnis/version-constraint/compare/dc5f5d1...main
+
+[#1]: https://github.com/ergebnis/version-constraint/pull/2
+
+[@localheinz]: https://github.com/localheinz